# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Aozora - Dorm Finder API for Lopez, Quezon
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # --- USERS ---
...
  /appointments:
    get:
      operationId: getAppointments
      tags: [appointments]
      summary: Get appointments for current user
      parameters:
        - name: role
          in: query
          schema:
            type: string
            enum: [student, owner]
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, approved, rejected]
      responses:
        "200":
          description: List of appointments
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppointmentsListResponse"

    post:
      operationId: createAppointment
      tags: [appointments]
      summary: Request a visitation appointment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAppointmentRequest"
      responses:
        "201":
          description: Appointment requested
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"
  /appointments/{appointmentId}:
    get:
      operationId: getAppointmentById
      tags: [appointments]
      summary: Get appointment details
      parameters:
        - name: appointmentId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Appointment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"

    put:
      operationId: updateAppointmentStatus
      tags: [appointments]
      summary: Approve or reject appointment (owner only)
      parameters:
        - name: appointmentId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAppointmentStatusRequest"
      responses:
        "200":
          description: Updated appointment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"

  # --- CONVERSATIONS & MESSAGES ---
...
  /admin/users:
    get:
      operationId: adminGetUsers
      tags: [admin]
      summary: List all users (admin only)
      parameters:
        - name: role
          in: query
          schema:
            type: string
        - name: verificationStatus
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
      responses:
        "200":
          description: Users list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUsersListResponse"
  /admin/users/{userId}/status:
    put:
      operationId: adminUpdateUserStatus
      tags: [admin]
      summary: Suspend or unsuspend a user (admin only)
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminUpdateUserStatusRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
...
  /admin/verifications/{verificationId}:
    put:
      operationId: adminReviewVerification
      tags: [admin]
      summary: Approve or reject a verification (admin only)
      parameters:
        - name: verificationId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminReviewVerificationRequest"
      responses:
        "200":
          description: Reviewed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VerificationRecord"
  /admin/dorms:
    get:
      operationId: adminGetDorms
      tags: [admin]
      summary: List all dorms for review (admin only)
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, approved, rejected]
      responses:
        "200":
          description: Dorms
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DormsListResponse"
  /admin/dorms/{dormId}/status:
    put:
      operationId: adminUpdateDormStatus
      tags: [admin]
      summary: Approve or reject a dorm listing (admin only)
      parameters:
        - name: dormId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminUpdateDormStatusRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Dorm"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required: [error]
...
    User:
      type: object
      properties:
        id:
          type: integer
        fullName:
          type: string
        email:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
        role:
          type: string
          enum: [student, owner, admin]
        verificationStatus:
          type: string
          enum: [unverified, pending, verified, rejected]
        isSuspended:
          type: boolean
        avatarUrl:
          type: string
          nullable: true
        createdAt:
          type: string
      required: [id, fullName, role, verificationStatus, isSuspended, createdAt]
    PublicUser:
      type: object
      properties:
        id:
          type: integer
        fullName:
          type: string
        role:
          type: string
        verificationStatus:
          type: string
        avatarUrl:
          type: string
          nullable: true
      required: [id, fullName, role, verificationStatus]
    RegisterRequest:
      type: object
      properties:
        fullName:
          type: string
        email:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
        password:
          type: string
        role:
          type: string
          enum: [student, owner]
      required: [fullName, password, role]
    LoginRequest:
      type: object
      properties:
        identifier:
          type: string
          description: Email or phone number
        password:
          type: string
      required: [identifier, password]
    AuthResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
      required: [token, user]
...
    VerificationRecord:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        idImageUrl:
          type: string
        idType:
          type: string
        status:
          type: string
          enum: [pending, approved, rejected]
        reviewNote:
          type: string
          nullable: true
        submittedAt:
          type: string
        reviewedAt:
          type: string
          nullable: true
      required: [id, userId, idImageUrl, idType, status, submittedAt]

    # --- DORM SCHEMAS ---
    Dorm:
      type: object
      properties:
        id:
          type: integer
        ownerId:
          type: integer
        name:
          type: string
        description:
          type: string
        monthlyRent:
          type: number
        address:
          type: string
        latitude:
          type: number
          nullable: true
        longitude:
          type: number
          nullable: true
        amenities:
          type: array
          items:
            type: string
        totalRooms:
          type: integer
        bedsPerRoom:
          type: integer
        availableBeds:
          type: integer
        status:
          type: string
          enum: [pending, approved, rejected]
        coverPhotoUrl:
          type: string
          nullable: true
        averageRating:
          type: number
          nullable: true
        totalReviews:
          type: integer
        createdAt:
          type: string
      required: [id, ownerId, name, description, monthlyRent, address, amenities, totalRooms, bedsPerRoom, availableBeds, status, totalReviews, createdAt]
...
    DormsListResponse:
      type: object
      properties:
        dorms:
          type: array
          items:
            $ref: "#/components/schemas/Dorm"
        total:
          type: integer
        page:
          type: integer
        totalPages:
          type: integer
      required: [dorms, total, page, totalPages]
...
    Appointment:
      type: object
      properties:
        id:
          type: integer
        studentId:
          type: integer
        dormId:
          type: integer
        preferredDate:
          type: string
        preferredTime:
          type: string
        message:
          type: string
          nullable: true
        status:
          type: string
          enum: [pending, approved, rejected]
        ownerNote:
          type: string
          nullable: true
        student:
          $ref: "#/components/schemas/PublicUser"
        dorm:
          $ref: "#/components/schemas/Dorm"
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, studentId, dormId, preferredDate, preferredTime, status, student, dorm, createdAt, updatedAt]
    AppointmentsListResponse:
      type: object
      properties:
        appointments:
          type: array
          items:
            $ref: "#/components/schemas/Appointment"
        total:
          type: integer
      required: [appointments, total]
    CreateAppointmentRequest:
      type: object
      properties:
        dormId:
          type: integer
        preferredDate:
          type: string
        preferredTime:
          type: string
        message:
          type: string
      required: [dormId, preferredDate, preferredTime]
    UpdateAppointmentStatusRequest:
      type: object
      properties:
        status:
          type: string
          enum: [approved, rejected]
        ownerNote:
          type: string
      required: [status]

    # --- MESSAGE SCHEMAS ---
...
    AdminUsersListResponse:
      type: object
      properties:
        users:
          type: array
          items:
            $ref: "#/components/schemas/User"
        total:
          type: integer
        page:
          type: integer
        totalPages:
          type: integer
      required: [users, total, page, totalPages]
    AdminUpdateUserStatusRequest:
      type: object
      properties:
        isSuspended:
          type: boolean
      required: [isSuspended]
...
    AdminReviewVerificationRequest:
      type: object
      properties:
        status:
          type: string
          enum: [approved, rejected]
        reviewNote:
          type: string
      required: [status]
    AdminUpdateDormStatusRequest:
      type: object
      properties:
        status:
          type: string
          enum: [approved, rejected]
        note:
          type: string
      required: [status]
...