# lib/api-zod/src/generated/types/adminUpdateUserStatusRequest.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Aozora - Dorm Finder API for Lopez, Quezon
 * OpenAPI spec version: 0.1.0
 */

export interface AdminUpdateUserStatusRequest {
  isSuspended: boolean;
}