# lib/api-zod/src/generated/types/dormsListResponse.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Aozora - Dorm Finder API for Lopez, Quezon
 * OpenAPI spec version: 0.1.0
 */
import type { Dorm } from './dorm';

export interface DormsListResponse {
  dorms: Dorm[];
  total: number;
  page: number;
  totalPages: number;
}