# lib/api-zod/src/generated/types/publicUser.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Aozora - Dorm Finder API for Lopez, Quezon
 * OpenAPI spec version: 0.1.0
 */

export interface PublicUser {
  id: number;
  fullName: string;
  role: string;
  verificationStatus: string;
  avatarUrl?: string | null;
}