# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Aozora - Dorm Finder API for Lopez, Quezon
 * OpenAPI spec version: 0.1.0
 */
import type { UserRole } from './userRole';
import type { UserVerificationStatus } from './userVerificationStatus';

export interface User {
  id: number;
  fullName: string;
  email?: string | null;
  phone?: string | null;
  role: UserRole;
  verificationStatus: UserVerificationStatus;
  isSuspended: boolean;
  avatarUrl?: string | null;
  createdAt: string;
}